apps/publisher/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
playwright==1.45.0
playwright-stealth==1.0.6
aioredis==2.0.1
httpx==0.27.0
cryptography==42.0.8
psycopg2-binary==2.9.9
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1
structlog==24.2.0
